Create Array[Easy4]_Set Matrix Zeroes
Logic: without using extra array space:
1) check the array elements one by one which contain 0 apply two for loops to check it.
2) here we are checking from i=0 and j=0 and the logic is making the respective  index i=0 index and j=0 index zero when matrix[i][j] caught 0 but this is not possible for 0th row and 0th column hence for them we will do bool expression true and make them 0 later this is one reason why we take for loop separately for 0th row and col.
3) Here we apply two for loops to make respective index of 0th element to be 0 we start them from zero as if we start from 0 then we can't do the 0 element index become 0 as the 0th index is 0 then it will make 0 to all according to loop and give wrong answer this is another reason  why we take for loop separately for 0th row and col.
4) Then we apply two for loops for making the row zero when column 0 is 0 and making column 0 when row 0 is 0. As, when firstRow got zero that row we will make zero in first condition only but its column we make zero in last loop , similarly for col. 


Leetcode logic:
The question seems to be pretty simple but the trick here is that we need to modify the given matrix in place i.e. our space complexity needs to O(1)O(1).
We will go through two different approaches to the question. The first approach makes use of additional memory while the other does not.
Approach 1: Additional Memory Approach
Intuition

If any cell of the matrix has a zero we can record its row and column number. All the cells of this recorded row and column can be marked zero in the next iteration.

Algorithm

We make a pass over our original array and look for zero entries.

If we find that an entry at [i, j] is 0, then we need to record somewhere the row i and column j.

So, we use two sets, one for the rows and one for the columns.

 if cell[i][j] == 0 {
     row_set.add(i)
     column_set.add(j)
 }
Finally, we iterate over the original matrix. For every cell we check if the row r or column c had been marked earlier. If any of them was marked, we set the value in the cell to 0.

 if r in row_set or c in column_set {
     cell[r][c] = 0
 }

Complexity Analysis

Time Complexity: O(M \times N)O(M×N) where M and N are the number of rows and columns respectively.
Space Complexity: O(M + N)O(M+N).

Approach 2: O(1) Space, Efficient Solution
Intuition

Rather than using additional variables to keep track of rows and columns to be reset, we use the matrix itself as the indicators.

The idea is that we can use the first cell of every row and column as a flag. This flag would determine whether a row or column has been set to zero. This means for every cell instead of going to M+NM+N cells and setting it to zero we just set the flag in two cells.

if cell[i][j] == 0 {
    cell[i][0] = 0
    cell[0][j] = 0
}
These flags are used later to update the matrix. If the first cell of a row is set to zero this means the row should be marked zero. If the first cell of a column is set to zero this means the column should be marked zero.

Algorithm

We iterate over the matrix and we mark the first cell of a row i and first cell of a column j, if the condition in the pseudo code above is satisfied. i.e. if cell[i][j] == 0.

The first cell of row and column for the first row and first column is the same i.e. cell[0][0]. Hence, we use an additional variable to tell us if the first column had been marked or not and the cell[0][0] would be used to tell the same for the first row.

Now, we iterate over the original matrix starting from second row and second column i.e. matrix[1][1] onwards. For every cell we check if the row r or column c had been marked earlier by checking the respective first row cell or first column cell. If any of them was marked, we set the value in the cell to 0. Note the first row and first column serve as the row_set and column_set that we used in the first approach.

We then check if cell[0][0] == 0, if this is the case, we mark the first row as zero.

And finally, we check if the first column was marked, we make all entries in it as zeros.

Complexity Analysis

Time Complexity : O(M \times N)O(M×N)
Space Complexity : O(1)O(1)