Create Array[Easy7]_Two Sum
Logic:
1) Find the difference between target and number in the array.
2) As if that difference present in the list it means that number pair is exist with which it form target.
3) We use hash table that is unordered mapping to check weather difference is present in the array or not/
4) We will store the pair of index and value one by one in the map if they are not present in the array .
5) Then we will check further elements and try to match the existing elements are matching with difference or not.
6) Once We find the difference we will return the array indexes.

Algorithm:
Using Hashmap

It is possible to solve this problem in linear time. The idea is to make use of a hashmap to store the indices of the elements that are already visited.  Hashmap "key" is the number in the given input array (You add this to the hashmap as you visit each element). Hashmap "value" is the index of the number in the array represented by the hashmap key.


For a given input array this algorithm does the following steps:

Create a hashmap which accepts integer datatype as key and value.

Iterate through each element in the given array starting from the first element.

In each iteration check if required number (required  number = target sum - current number) is present in the hashmap.

If present, return {required number index, current number index} as  result.

Otherwise add the current iteration number as key and its index as value to the hashmap. Repeat this  until you find the result.