Create Array_Cyclically rotate an array by one
1) Method 1:
1) Store last element in a variable say x. 
2) Shift all elements one position ahead. 
3) Replace first element of array with x.

2)Method 2:
We can use two pointers, say i and j which point to first and last element of array respectively. 
As we know in cyclic rotation we will bring last element to first and shift rest in forward direction, 
so start swaping arr[i] and arr[j] and keep j fixed and i moving towards j.  Repeat till i is not equal to j.